# docs/api/control.openapi.json
{
  "info": {
    "title": "one-fetch Control API",
    "version": "0.1.0",
    "description": "Canonical one-fetch Control API. Runtime capabilities remain authoritative."
  },
  "openapi": "3.1.0",
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque administrator access token"
      },
      "executionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque execution token scoped to its own report"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/tokens/execution": {
      "get": {
        "security": [{ "adminBearer": [] }],
        "responses": {
          "200": {
            "description": "Execution tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schemaVersion": { "type": "number", "enum": [1] },
                    "tokens": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "schemaVersion": { "type": "number", "enum": [1] },
                          "id": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128,
                            "pattern": "^[A-Za-z0-9._:-]+$"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 128
                          },
                          "scope": {
                            "type": "object",
                            "properties": {
                              "transports": {
                                "type": "array",
                                "items": {
                                  "type": "string",
                                  "enum": ["http", "websocket", "tcp", "tls"]
                                },
                                "minItems": 1,
                                "maxItems": 4
                              },
                              "origins": {
                                "type": "array",
                                "items": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 2048
                                },
                                "maxItems": 256
                              },
                              "ports": {
                                "type": "array",
                                "items": {
                                  "type": "integer",
                                  "minimum": 1,
                                  "maximum": 65535
                                },
                                "maxItems": 256
                              }
                            },
                            "required": ["transports", "origins", "ports"],
                            "additionalProperties": false
                          },
                          "quota": {
                            "type": "object",
                            "properties": {
                              "requestsPerMinute": {
                                "type": "integer",
                                "exclusiveMinimum": 0,
                                "maximum": 1000000
                              },
                              "burst": {
                                "type": "integer",
                                "exclusiveMinimum": 0,
                                "maximum": 1000000
                              },
                              "concurrentHttp": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 100000
                              },
                              "concurrentTunnels": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 100000
                              },
                              "bytesPerDay": {
                                "type": "integer",
                                "exclusiveMinimum": 0,
                                "maximum": 9007199254740991
                              }
                            },
                            "required": [
                              "requestsPerMinute",
                              "burst",
                              "concurrentHttp",
                              "concurrentTunnels",
                              "bytesPerDay"
                            ],
                            "additionalProperties": false
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "revokedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "schemaVersion",
                          "id",
                          "name",
                          "scope",
                          "quota",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      },
                      "maxItems": 10000
                    }
                  },
                  "required": ["schemaVersion", "tokens"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "adminBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "schemaVersion": { "type": "number", "enum": [1] },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "scope": {
                    "type": "object",
                    "properties": {
                      "transports": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": ["http", "websocket", "tcp", "tls"]
                        },
                        "minItems": 1,
                        "maxItems": 4
                      },
                      "origins": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2048
                        },
                        "maxItems": 256
                      },
                      "ports": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 65535
                        },
                        "maxItems": 256
                      }
                    },
                    "required": ["transports", "origins", "ports"],
                    "additionalProperties": false
                  },
                  "quota": {
                    "type": "object",
                    "properties": {
                      "requestsPerMinute": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 1000000
                      },
                      "burst": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 1000000
                      },
                      "concurrentHttp": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100000
                      },
                      "concurrentTunnels": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100000
                      },
                      "bytesPerDay": {
                        "type": "integer",
                        "exclusiveMinimum": 0,
                        "maximum": 9007199254740991
                      }
                    },
                    "required": [
                      "requestsPerMinute",
                      "burst",
                      "concurrentHttp",
                      "concurrentTunnels",
                      "bytesPerDay"
                    ],
                    "additionalProperties": false
                  },
                  "expiresAt": { "type": "string", "format": "date-time" }
                },
                "required": ["schemaVersion", "name", "scope", "quota"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schemaVersion": { "type": "number", "enum": [1] },
                    "credential": {
                      "type": "object",
                      "properties": {
                        "schemaVersion": { "type": "number", "enum": [1] },
                        "id": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "scope": {
                          "type": "object",
                          "properties": {
                            "transports": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "enum": ["http", "websocket", "tcp", "tls"]
                              },
                              "minItems": 1,
                              "maxItems": 4
                            },
                            "origins": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 2048
                              },
                              "maxItems": 256
                            },
                            "ports": {
                              "type": "array",
                              "items": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 65535
                              },
                              "maxItems": 256
                            }
                          },
                          "required": ["transports", "origins", "ports"],
                          "additionalProperties": false
                        },
                        "quota": {
                          "type": "object",
                          "properties": {
                            "requestsPerMinute": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 1000000
                            },
                            "burst": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 1000000
                            },
                            "concurrentHttp": {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 100000
                            },
                            "concurrentTunnels": {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 100000
                            },
                            "bytesPerDay": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            }
                          },
                          "required": [
                            "requestsPerMinute",
                            "burst",
                            "concurrentHttp",
                            "concurrentTunnels",
                            "bytesPerDay"
                          ],
                          "additionalProperties": false
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "revokedAt": { "type": "string", "format": "date-time" }
                      },
                      "required": [
                        "schemaVersion",
                        "id",
                        "name",
                        "scope",
                        "quota",
                        "createdAt"
                      ],
                      "additionalProperties": false
                    },
                    "token": {
                      "type": "string",
                      "minLength": 32,
                      "maxLength": 4096
                    }
                  },
                  "required": ["schemaVersion", "credential", "token"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/execution/{tokenId}": {
      "delete": {
        "security": [{ "adminBearer": [] }],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9._:-]+$"
            },
            "required": true,
            "name": "tokenId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked execution token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schemaVersion": { "type": "number", "enum": [1] },
                    "id": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128,
                      "pattern": "^[A-Za-z0-9._:-]+$"
                    },
                    "revokedAt": { "type": "string", "format": "date-time" }
                  },
                  "required": ["schemaVersion", "id", "revokedAt"],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Token not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[a-z0-9_]+$"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1024
                        },
                        "retryable": { "type": "boolean" },
                        "correlationId": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128,
                          "pattern": "^[A-Za-z0-9._:-]+$"
                        }
                      },
                      "required": ["code", "message"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "servers": [
    { "url": ".", "description": "The configured Control service base URL" }
  ],
...
}